:truck: Move Resources out of Renderer
