+++ src/modules/game/resources/index.ts
@@ -1,7 +1,7 @@
 import { initLoader, runLoader } from './loader'
-import { IResources, Sheet } from '../../typings'
+import { IResources, Sheet } from '../typings'
 import { LoaderResource } from 'pixi.js-legacy'
-import { SPRITE_JSON_PATH } from '../../config/sprite'
+import { SPRITE_JSON_PATH } from '../config/sprite'
 
 /** A manager for PIXI resources. */
 export default class Resources implements IResources {
